REGRESSION (r199401): Internal builds of Safari hang on launch https://bugs.webkit.org/show_bug.cgi?id=156545 <rdar://problem/25697779>
Reviewed by Anders Carlsson.

For some reason SecCodeCopyGuestWithAttributes() is failing with an error in Apple Internal
Safari builds. For now, temporarily allow the failure while I investigate the cause in
<rdar://problem/25706517>.

* Shared/mac/CodeSigning.mm:
(WebKit::secCodeForProcess): Log the failure with OSStatus code and return nullptr;
(WebKit::codeSigningIdentifierForProcess): Return a null string if secCodeForProcess() returns a nullptr.
This will cause us to treat affected Apple Internal Safari builds the same as we would treat
an unsigned or third-party signed app.

Identifier: 174652@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199504 268f45cc-cd09-0410-ab3c-d52691b4dbfc