[GTK] Test /webkit2/WebKitWebView/mouse-target fails with overlay scrollbars https://bugs.webkit.org/show_bug.cgi?id=156621
Reviewed by Martin Robinson.

This is a problem of the test itself, we are using an overflow that is not actually scrollable, but that shows
scrollbar troughs when not using overlay scrollbars because it's marked as scrollable. It would be easier to
simply use the main frame scrollbars to test that and it will work no matter of scrollbars are overlay or not.

* TestWebKitAPI/Tests/WebKit2Gtk/TestUIClient.cpp:
(testWebViewMouseTarget): Remove the div and move the mouse over the main frame scrollbar to test scrolbars hit test.

Identifier: 174788@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199653 268f45cc-cd09-0410-ab3c-d52691b4dbfc