buildObjectForEventListener should not call into JSC with a null ExecState https://bugs.webkit.org/show_bug.cgi?id=156923
Reviewed by Joseph Pecoraro.

If a user had disabled JavaScript on their page then the inspector tried to
add an event listener we would fail to create an ExecState. Since we didn't
check this ExecState was valid we would then attempt to stringify the value,
which would cause JSC to crash.

* inspector/InspectorDOMAgent.cpp:
(WebCore::InspectorDOMAgent::buildObjectForEventListener):


Identifier: 175031@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199905 268f45cc-cd09-0410-ab3c-d52691b4dbfc