Minor refactoring in RenderMathMLOperator https://bugs.webkit.org/show_bug.cgi?id=156906
Patch by Frederic Wang <fwang@igalia.com> on 2016-04-25
Reviewed by Martin Robinson.

No new tests, this is only minor refactoring that does not change the behavior.

* rendering/mathml/RenderMathMLOperator.cpp:
(WebCore::RenderMathMLOperator::getGlyphAssemblyFallBack):
We rename the "state" integer to an "expected" enum indicating the next expected part.
(WebCore::RenderMathMLOperator::paintGlyph): We add a missing dot at the end of a sequence.
We also replace ceil(x+1) with ceil(x)+1 to get rid of the temporary variable.

Identifier: 175080@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@199978 268f45cc-cd09-0410-ab3c-d52691b4dbfc